Resolve stats dashboard tags issue and refactor dashboard FXML
The tags displayed on the statistics dashboard did not follow the UI convention due to its style being hardcoded.
Additionally, the code for the Statistics dashboard has the FXMl and styles all defined in the Java class/controller.

Resolving the tag issue is necessary to keep the UI consistent. Additionally, the FXML style and node code being in the class
violates separation of concerns.

Lets resolve the tag issue and improve quality of code by refactoring the FXML component and design code into its FXML file.